Create a new job.

Create a new job.

Usage

create_job(
  client,
  access_control_list = NULL,
  compute = NULL,
  continuous = NULL,
  deployment = NULL,
  description = NULL,
  edit_mode = NULL,
  email_notifications = NULL,
  format = NULL,
  git_source = NULL,
  health = NULL,
  job_clusters = NULL,
  max_concurrent_runs = NULL,
  name = NULL,
  notification_settings = NULL,
  parameters = NULL,
  queue = NULL,
  run_as = NULL,
  schedule = NULL,
  tags = NULL,
  tasks = NULL,
  timeout_seconds = NULL,
  trigger = NULL,
  webhook_notifications = NULL
)

jobsCreate(
  client,
  access_control_list = NULL,
  compute = NULL,
  continuous = NULL,
  deployment = NULL,
  description = NULL,
  edit_mode = NULL,
  email_notifications = NULL,
  format = NULL,
  git_source = NULL,
  health = NULL,
  job_clusters = NULL,
  max_concurrent_runs = NULL,
  name = NULL,
  notification_settings = NULL,
  parameters = NULL,
  queue = NULL,
  run_as = NULL,
  schedule = NULL,
  tags = NULL,
  tasks = NULL,
  timeout_seconds = NULL,
  trigger = NULL,
  webhook_notifications = NULL
)

Arguments

client

Required. Instance of DatabricksClient()

access_control_list

List of permissions to set on the job.

compute

A list of compute requirements that can be referenced by tasks of this job.

continuous

An optional continuous property for this job.

deployment

Deployment information for jobs managed by external sources.

description

An optional description for the job.

edit_mode

Edit mode of the job.

email_notifications

An optional set of email addresses that is notified when runs of this job begin or complete as well as when this job is deleted.

format

Used to tell what is the format of the job.

git_source

An optional specification for a remote Git repository containing the source code used by tasks.

health

An optional set of health rules that can be defined for this job.

job_clusters

A list of job cluster specifications that can be shared and reused by tasks of this job.

max_concurrent_runs

An optional maximum allowed number of concurrent runs of the job.

name

An optional name for the job.

notification_settings

Optional notification settings that are used when sending notifications to each of the email_notifications and webhook_notifications for this job.

parameters

Job-level parameter definitions.

queue

The queue settings of the job.

run_as

Write-only setting, available only in Create/Update/Reset and Submit calls.

schedule

An optional periodic schedule for this job.

tags

A map of tags associated with the job.

tasks

A list of task specifications to be executed by this job.

timeout_seconds

An optional timeout applied to each run of this job.

trigger

A configuration to trigger a run when certain conditions are met.

webhook_notifications

A collection of system notification IDs to notify when runs of this job begin or complete.