Update access list.

Updates an existing IP access list, specified by its ID.

Usage

ipAccessListsUpdate(
  client,
  ip_access_list_id,
  enabled = NULL,
  ip_addresses = NULL,
  label = NULL,
  list_type = NULL
)

Arguments

client

Required. Instance of DatabricksClient()

ip_access_list_id

Required. The ID for the corresponding IP access list.

enabled

Specifies whether this IP access list is enabled.

ip_addresses

label

Label for the IP access list.

list_type

Type of IP access list.

Details

A list can include allow lists and block lists. See the top of this file for a description of how the server treats allow lists and block lists at run time.

When updating an IP access list:

For all allow lists and block lists combined, the API supports a maximum of 1000 IP/CIDR values, where one CIDR counts as a single value. Attempts to exceed that number return error 400 with error_code value QUOTA_EXCEEDED.

If the updated list would block the calling user's current IP, error 400 is returned with error_code value INVALID_STATE.

It can take a few minutes for the changes to take effect. Note that your resulting IP access list has no effect until you enable the feature. See :method:workspaceconf/setStatus.