Validate a storage credential.

Validates a storage credential. At least one of external_location_name and url need to be provided. If only one of them is provided, it will be used for validation. And if both are provided, the url will be used for validation, and external_location_name will be ignored when checking overlapping urls.

Usage

validate_storage_credential(
  client,
  aws_iam_role = NULL,
  azure_managed_identity = NULL,
  azure_service_principal = NULL,
  cloudflare_api_token = NULL,
  databricks_gcp_service_account = NULL,
  external_location_name = NULL,
  read_only = NULL,
  storage_credential_name = NULL,
  url = NULL
)

storageCredentialsValidate(
  client,
  aws_iam_role = NULL,
  azure_managed_identity = NULL,
  azure_service_principal = NULL,
  cloudflare_api_token = NULL,
  databricks_gcp_service_account = NULL,
  external_location_name = NULL,
  read_only = NULL,
  storage_credential_name = NULL,
  url = NULL
)

Arguments

client

Required. Instance of DatabricksClient()

aws_iam_role

The AWS IAM role configuration.

azure_managed_identity

The Azure managed identity configuration.

azure_service_principal

The Azure service principal configuration.

cloudflare_api_token

The Cloudflare API token configuration.

databricks_gcp_service_account

The Databricks created GCP service account configuration.

external_location_name

The name of an existing external location to validate.

read_only

Whether the storage credential is only usable for read operations.

storage_credential_name

The name of the storage credential to validate.

url

The external location url to validate.

Details

Either the storage_credential_name or the cloud-specific credential must be provided.

The caller must be a metastore admin or the storage credential owner or have the CREATE_EXTERNAL_LOCATION privilege on the metastore and the storage credential.